% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_clustering.R
\name{supervised_pam}
\alias{supervised_pam}
\title{Algoritme PAM supervisat}
\usage{
supervised_pam(datos, clases, k_por_clase = 1, lambda = 0)
}
\arguments{
\item{datos}{Matriu o data.frame de mostres.}

\item{clases}{Vector de classes.}

\item{k_por_clase}{Nombre de prototips per classe.}

\item{lambda}{No utilitzat actualment.}
}
\value{
Llista amb prototips, classes, prediccions, matriu de confusió i precisió.
}
\description{
Aplica el mètode PAM de forma supervisada seleccionant prototips per classe.
}
