% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_clustering.R
\name{supervised_sridhcr}
\alias{supervised_sridhcr}
\title{Algoritmo SRIDHCR supervisado}
\usage{
supervised_sridhcr(datos, clases, beta = 0.4, max_iter = 50, reinicios = 5)
}
\arguments{
\item{datos}{Matriz o data.frame de muestras (se estandariza internamente).}

\item{clases}{Vector de clases (factor).}

\item{beta}{Peso de la penalización por número de clusters.}

\item{max_iter}{Número máximo de iteraciones por reinicio.}

\item{reinicios}{Número de reinicios aleatorios para evitar mínimos locales.}
}
\value{
Una lista con prototipos, predicciones, matriz de confusión, impureza, precisión y valor de la función q.
}
\description{
Implementa el algoritmo de clustering supervisado SRIDHCR (Single Representative Insertion/Deletion Hill Climbing with Random Restart)
según lo descrito en el artículo de Eick et al. (2004).
}
