% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_clustering.R
\name{supervised_pam}
\alias{supervised_pam}
\title{Algoritmo PAM supervisado}
\usage{
supervised_pam(datos, clases, k = 2, beta = 1, max_iter = 50)
}
\arguments{
\item{datos}{Matriz o data.frame de muestras.}

\item{clases}{Vector de clases.}

\item{k}{Número total de prototipos a seleccionar.}

\item{beta}{Peso de la penalización.}

\item{max_iter}{Número máximo de iteraciones para la fase de mejora.}
}
\value{
Lista con prototipos, predicciones, impureza, coste y matriz de confusión.
}
\description{
Aplica el método PAM de forma supervisada seleccionando prototipos globales que minimizan la impureza y penalización.
según lo descrito en el artículo de Eick et al. (2004).
}
