% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supervised_clustering.R
\name{supervised_sridhcr}
\alias{supervised_sridhcr}
\title{Algoritme SRIDHCR supervisat}
\usage{
supervised_sridhcr(datos, clases, max_iter = 20, max_reinicios = 5)
}
\arguments{
\item{datos}{Matriu o data.frame de mostres.}

\item{clases}{Vector de classes.}

\item{max_iter}{Nombre màxim d'iteracions per reinici.}

\item{max_reinicios}{Nombre màxim de reinicis aleatoris.}
}
\value{
Millor resultat obtingut (prototips, prediccions, precisió, etc.).
}
\description{
Aplica l'algoritme SRIDHCR mitjançant optimització iterativa.
}
